Dedupe the helper file
=============================== Coverage summary ================
Statements   : 89.31% ( 11244/12590 ), 12 ignored
Branches     : 83.34% ( 5003/6003 ), 71 ignored
Functions    : 91.27% ( 836/916 ), 11 ignored
Lines        : 89.86% ( 10009/11139 )
=================================================================
