+++ src/pre/bounty.js
@@ -2,7 +2,7 @@ import {
   ASSERT,
   TRACE,
   THROW,
-} from './helpers';
+} from '../helpers';
 
 import {
   ML_START,
