+++ src/pre/cutter.js
@@ -6,7 +6,7 @@ import {
   ASSERT,
   TRACE,
   THROW,
-} from './helpers';
+} from '../helpers';
 
 import {
   ML_START,
