+++ src/pre/ml2dsl.js
@@ -8,7 +8,7 @@ import {
   ASSERT,
   TRACE,
   THROW,
-} from './helpers';
+} from '../helpers';
 import {
   domain__debug,
   domain_getValue,
