+++ src/pre/runner.js
@@ -15,7 +15,7 @@ import {
   ASSERT_NORDOM,
   TRACE,
   THROW,
-} from './helpers';
+} from '../helpers';
 import {
   ml_countConstraints,
   ml_getOpList,
